requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
packaging==23.2
python-multipart==0.0.6
gitpython==3.1.42 